requirements.txt
Django
django-cors-headers
djangorestframework
requests
requests-oauthlib